Upgrade selection_context
